requirements.txt
PySensors >= 0.0.4
pyserial >= 3.5
PyYAML >= 6.0.0
